Refactor utilities import path in app.py
The import path for utilities in the app.py has been reorganized in this commit. This change was necessary due to a rename operation which moved the utilities from the local directory to a more global scripts directory.
